Refactor AddProject page to use DBClass methods
Updated AddProject.cshtml to display user names as FirstName LastName.
Refactored AddProject.cshtml.cs to use DBClass for data operations,
removing hardcoded connection strings and private methods.
Updated DBClass.cs to add GrantModel and implement methods for
loading faculty members, grants, and adding projects.
